services/chat/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
openai==0.28.1
requests==2.31.0
